# api/manager/swagger.yaml
basePath: /
definitions:
...
  d7y_io_dragonfly_v2_manager_models.Config:
    properties:
      bio:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_del:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.User'
      user_id:
        type: integer
      value:
        type: string
    type: object
  d7y_io_dragonfly_v2_manager_models.JSONMap:
    additionalProperties: {}
    type: object
  d7y_io_dragonfly_v2_manager_models.Job:
    properties:
      args:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.JSONMap'
      bio:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_del:
        type: integer
      result:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.JSONMap'
      scheduler_clusters:
        items:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.SchedulerCluster'
        type: array
      seed_peer_clusters:
        items:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.SeedPeerCluster'
        type: array
      state:
        type: string
      task_id:
        type: string
      type:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.User'
      user_id:
        type: integer
    type: object
...
  d7y_io_dragonfly_v2_manager_models.Peer:
    properties:
      build_platform:
        type: string
      created_at:
        type: string
      download_port:
        type: integer
      git_commit:
        type: string
      git_version:
        type: string
      host_name:
        type: string
      id:
        type: integer
      idc:
        type: string
      ip:
        type: string
      is_del:
        type: integer
      kernel_version:
        type: string
      location:
        type: string
      object_storage_port:
        type: integer
      os:
        type: string
      platform:
        type: string
      platform_family:
        type: string
      platform_version:
        type: string
      port:
        type: integer
      proxy_port:
        type: integer
      scheduler_cluster:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.SchedulerCluster'
      scheduler_cluster_id:
        type: integer
      state:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  d7y_io_dragonfly_v2_manager_models.Scheduler:
    properties:
      config:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.JSONMap'
      created_at:
        type: string
      features:
        items:
          type: string
        type: array
      host_name:
        type: string
      id:
        type: integer
      idc:
        type: string
      ip:
        type: string
      is_del:
        type: integer
      last_keep_alive_at:
        type: string
      location:
        type: string
      port:
        type: integer
      scheduler_cluster:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.SchedulerCluster'
      scheduler_cluster_id:
        type: integer
      state:
        type: string
      updated_at:
        type: string
    type: object
  d7y_io_dragonfly_v2_manager_models.SchedulerCluster:
    properties:
      bio:
        type: string
      client_config:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.JSONMap'
      config:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.JSONMap'
      created_at:
        type: string
      id:
        type: integer
      is_default:
        type: boolean
      is_del:
        type: integer
      jobs:
        items:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.Job'
        type: array
      name:
        type: string
      peers:
        items:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.Peer'
        type: array
      schedulers:
        items:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.Scheduler'
        type: array
      scopes:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.JSONMap'
      seed_peer_clusters:
        items:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.SeedPeerCluster'
        type: array
      updated_at:
        type: string
    type: object
  d7y_io_dragonfly_v2_manager_models.SeedPeer:
    properties:
      created_at:
        type: string
      download_port:
        type: integer
      host_name:
        type: string
      id:
        type: integer
      idc:
        type: string
      ip:
        type: string
      is_del:
        type: integer
      location:
        type: string
      object_storage_port:
        type: integer
      port:
        type: integer
      seed_peer_cluster:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.SeedPeerCluster'
      seed_peer_cluster_id:
        type: integer
      state:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  d7y_io_dragonfly_v2_manager_models.SeedPeerCluster:
    properties:
      bio:
        type: string
      config:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.JSONMap'
      created_at:
        type: string
      id:
        type: integer
      is_del:
        type: integer
      jobs:
        items:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.Job'
        type: array
      name:
        type: string
      scheduler_clusters:
        items:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.SchedulerCluster'
        type: array
      seed_peer:
        items:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.SeedPeer'
        type: array
      updated_at:
        type: string
    type: object
  d7y_io_dragonfly_v2_manager_models.User:
    properties:
      avatar:
        type: string
      bio:
        type: string
      configs:
        items:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.Config'
        type: array
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_del:
        type: integer
      location:
        type: string
      name:
        type: string
      phone:
        type: string
      state:
        type: string
      updated_at:
        type: string
    type: object
  d7y_io_dragonfly_v2_manager_permission_rbac.Permission:
    properties:
      action:
        enum:
        - read
        - '*'
        type: string
      object:
        type: string
    required:
    - action
    - object
    type: object
  d7y_io_dragonfly_v2_manager_types.AddPermissionForRoleRequest:
    properties:
      action:
        enum:
        - read
        - '*'
        type: string
      object:
        type: string
    required:
    - action
    - object
    type: object
...
  d7y_io_dragonfly_v2_manager_types.CreatePeerRequest:
    properties:
      build_platform:
        type: string
      download_port:
        type: integer
      git_commit:
        type: string
      git_version:
        type: string
      host_name:
        type: string
      idc:
        type: string
      ip:
        type: string
      kernel_version:
        type: string
      location:
        type: string
      object_storage_port:
        type: integer
      os:
        type: string
      platform:
        type: string
      platform_family:
        type: string
      platform_version:
        type: string
      port:
        type: integer
      proxy_port:
        type: integer
      scheduler_cluster_id:
        type: integer
      state:
        enum:
        - active
        - inactive
        type: string
      type:
        enum:
        - super
        - normal
        type: string
    required:
    - download_port
    - host_name
    - ip
    - port
    - proxy_port
    - scheduler_cluster_id
    - type
    type: object
...
  d7y_io_dragonfly_v2_manager_types.CreateRoleRequest:
    properties:
      permissions:
        items:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_permission_rbac.Permission'
        type: array
      role:
        type: string
    required:
    - permissions
    - role
    type: object
...
  d7y_io_dragonfly_v2_manager_types.CreateSchedulerRequest:
    properties:
      config:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.SchedulerConfig'
      features:
        items:
          type: string
        type: array
      host_name:
        type: string
      idc:
        type: string
      ip:
        type: string
      location:
        type: string
      port:
        type: integer
      scheduler_cluster_id:
        type: integer
    required:
    - host_name
    - ip
    - port
    - scheduler_cluster_id
    type: object
...
  d7y_io_dragonfly_v2_manager_types.DeletePermissionForRoleRequest:
    properties:
      action:
        enum:
        - read
        - '*'
        type: string
      object:
        type: string
    required:
    - action
    - object
    type: object
...
  d7y_io_dragonfly_v2_manager_types.SchedulerConfig:
    properties:
      manager_keep_alive_interval:
        $ref: '#/definitions/time.Duration'
    required:
    - manager_keep_alive_interval
    type: object
...
  d7y_io_dragonfly_v2_manager_types.UpdateSchedulerRequest:
    properties:
      config:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.SchedulerConfig'
      features:
        items:
          type: string
        type: array
      idc:
        type: string
      ip:
        type: string
      location:
        type: string
      port:
        type: integer
      scheduler_cluster_id:
        type: integer
      scheduler_id:
        type: integer
    type: object
...
  time.Duration:
    enum:
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    type: integer
    x-enum-varnames:
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
host: localhost:8080
info:
  contact:
    url: https://d7y.io
  description: Dragonfly Manager Server
  license:
    name: Apache 2.0
  title: Dragonfly Manager
  version: 1.0.0
paths:
...
  /api/v1/peers:
    get:
      consumes:
      - application/json
      description: Get Peers
      parameters:
      - default: 0
        description: current page
        in: query
        name: page
        required: true
        type: integer
      - default: 10
        description: return max item count, default 10, max 50
        in: query
        maximum: 50
        minimum: 2
        name: per_page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.Peer'
            type: array
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get Peers
      tags:
      - Peer
    post:
      consumes:
      - application/json
      description: Create by json config
      parameters:
      - description: Peer
        in: body
        name: Peer
        required: true
        schema:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.CreatePeerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.Peer'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Create Peer
      tags:
      - Peer
  /api/v1/peers/{id}:
    delete:
      consumes:
      - application/json
      description: Destroy by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Destroy Peer
      tags:
      - Peer
    get:
      consumes:
      - application/json
      description: Get Peer by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.Peer'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get Peer
      tags:
      - Peer
...
  /api/v1/roles:
    get:
      consumes:
      - application/json
      description: Get roles
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get Roles
      tags:
      - Role
    post:
      consumes:
      - application/json
      description: Create Role by json config
      parameters:
      - description: Role
        in: body
        name: Role
        required: true
        schema:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Create Role
      tags:
      - Role
  /api/v1/roles/{role}:
    delete:
      consumes:
      - application/json
      description: Destroy role by json config
      parameters:
      - description: role
        in: path
        name: role
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Destroy Role
      tags:
      - Role
    get:
      consumes:
      - application/json
      description: Get Role
      parameters:
      - description: role
        in: path
        name: role
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get Role
      tags:
      - Role
  /api/v1/roles/{role}/permissions:
    delete:
      consumes:
      - application/json
      description: Remove Role Permission by json config
      parameters:
      - description: Permission
        in: body
        name: Permission
        required: true
        schema:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.DeletePermissionForRoleRequest'
      - description: role
        in: path
        name: role
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Update Role
      tags:
      - Role
    post:
      consumes:
      - application/json
      description: Add Permission by json config
      parameters:
      - description: Permission
        in: body
        name: Permission
        required: true
        schema:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.AddPermissionForRoleRequest'
      - description: role
        in: path
        name: role
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Add Permission For Role
      tags:
      - Role
...
  /api/v1/schedulers:
    get:
      consumes:
      - application/json
      description: Get Schedulers
      parameters:
      - default: 0
        description: current page
        in: query
        name: page
        required: true
        type: integer
      - default: 10
        description: return max item count, default 10, max 50
        in: query
        maximum: 50
        minimum: 2
        name: per_page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.Scheduler'
            type: array
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get Schedulers
      tags:
      - Scheduler
    post:
      consumes:
      - application/json
      description: Create by json config
      parameters:
      - description: Scheduler
        in: body
        name: Scheduler
        required: true
        schema:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.CreateSchedulerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.Scheduler'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Create Scheduler
      tags:
      - Scheduler
  /api/v1/schedulers/{id}:
    delete:
      consumes:
      - application/json
      description: Destroy by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Destroy Scheduler
      tags:
      - Scheduler
    get:
      consumes:
      - application/json
      description: Get Scheduler by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.Scheduler'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get Scheduler
      tags:
      - Scheduler
    patch:
      consumes:
      - application/json
      description: Update by json config
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: Scheduler
        in: body
        name: Scheduler
        required: true
        schema:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.UpdateSchedulerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.Scheduler'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update Scheduler
      tags:
      - Scheduler
...
swagger: "2.0"
...